+++ src/api/put_object_lock_configuration.rs
@@ -61,7 +61,7 @@ impl<'a> PutObjectLockConfigurationFluentBuilder<'a> {
         self
     }
 
-    pub fn build_request(&self) -> Result<S3Request, Error> {
+    pub fn build_request(&self, now: std::time::SystemTime) -> Result<S3Request, Error> {
         let bucket = required(self.bucket.as_deref(), "bucket")?;
 
         let xml_body = build_object_lock_configuration_xml(&self.object_lock_configuration);
@@ -83,7 +83,7 @@ impl<'a> PutObjectLockConfigurationFluentBuilder<'a> {
             extra_headers.push((header_name, &computed_checksum));
         }
 
-        Ok(build_signed_request(
+        build_signed_request(
             &self.client.config_ref(),
             "PUT",
             bucket,
@@ -91,7 +91,8 @@ impl<'a> PutObjectLockConfigurationFluentBuilder<'a> {
             &extra_headers,
             xml_body.as_bytes(),
             Some(&[("object-lock", "")]),
-        ))
+            now,
+        )
     }
 
     pub fn parse_response(